style(app): extracted components to top level directory
also rearranged the fonts
