+++ app/layout.tsx
@@ -1,5 +1,5 @@
-import Footer from "@/app/ui/footer";
-import Header from "@/app/ui/header";
+import Footer from "@/components/footer";
+import Header from "@/components/header";
 import "@/styles/globals.css";
 import { config } from "@fortawesome/fontawesome-svg-core";
 import "@fortawesome/fontawesome-svg-core/styles.css";
@@ -11,12 +11,12 @@ import localFont from "next/font/local";
 config.autoAddCss = false;
 
 const geistSans = localFont({
-  src: "../public/fonts/GeistVF.woff",
+  src: "../fonts/GeistVF.woff",
   variable: "--font-geist-sans",
   weight: "100 900",
 });
 const geistMono = localFont({
-  src: "../public/fonts/GeistMonoVF.woff",
+  src: "../fonts/GeistMonoVF.woff",
   variable: "--font-geist-mono",
   weight: "100 900",
 });
